requirements.txt
atlassian-python-api==3.32.1
certifi==2022.12.7
charset-normalizer==2.1.1
Deprecated==1.2.13
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
urllib3==1.26.13
wrapt==1.14.1
